app/requirements.txt
Django==2.2.6
djangorestframework==3.10.3
markdown==3.1.1
django-filter==2.2.0
gunicorn==19.9.0
psycopg2-binary==2.8.3
django-rest-auth==0.9.5
django-allauth==0.40.0
django-rest-swagger==2.2.0
